// frontend/src/app/model/paciente.ts
export class Paciente {
    nomeCompleto: string = '';
    nomeSocial: string = '';
    nomeMae: string = '';
    nomePai: string = '';
    dataNascimento: string = '';
    sexo: string = '';
    nacionalidade: string = '';
    municipioNascimento: string = '';
    racaCor: string = '';
    frequentaEscola: string = '';
    escolaridade: string = '';
    situacaoFamiliar: string = '';
    vinculo: string = '';
    cadastro: string = '';
    deficiencia: string = '';
    celular: string = '';
    residencial: string = '';
    comercial: string = '';
    contato: string = '';
    email: string = '';
    enderecoOrigem: string = '';
    endereco: string = '';
    cpf: string = '';
    identidade: string = '';
  }
  